# api/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenEdusync API",
        "description": "API documentation for the OpenEdusy application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:\/\/localhost\/api\/v1",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "\/api\/CurriculumDetail": {
            "get": {
                "tags": [
                    "CurriculumDetail"
                ],
                "summary": "Get paginated list of CurriculumDetail",
                "description": "Retrieve a paginated list of CurriculumDetail with optional search",
                "operationId": "getCurriculumDetailPaginated",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "filter[curriculum_id]",
                        "in": "query",
                        "description": "Filter by curriculum ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Paginate the results",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/PaginatedCurriculumDetailResponse200"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CurriculumDetail"
                ],
                "summary": "Create a new CurriculumDetail",
                "description": " Create a new CurriculumDetail with the provided details",
                "operationId": "createCurriculumDetail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/CurriculumDetail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CurriculumDetail created successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/CreateCurriculumDetailResponse200"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/api\/CurriculumDetail\/{id}": {
            "get": {
                "tags": [
                    "CurriculumDetail"
                ],
                "summary": "Get a specific CurriculumDetail",
                "description": "Retrieve a CurriculumDetail by its ID",
                "operationId": "getCurriculumDetailById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/GetCurriculumDetailResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CurriculumDetail not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "CurriculumDetail"
                ],
                "summary": "Update a CurriculumDetail",
                "description": "Update an existing CurriculumDetail with the provided details",
                "operationId": "updateCurriculumDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/CurriculumDetail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CurriculumDetail updated successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/UpdateCurriculumDetailResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CurriculumDetail not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CurriculumDetail"
                ],
                "summary": "Delete a CurriculumDetail",
                "description": "Delete a CurriculumDetail by its ID",
                "operationId": "deleteCurriculumDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "CurriculumDetail deleted successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/DeleteCurriculumDetailResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CurriculumDetail not found"
                    },
                    "500": {
                        "description": " Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/api\/CurriculumDetail\/multiple": {
            "post": {
                "tags": [
                    "CurriculumDetail"
                ],
                "summary": "Create multiple CurriculumDetail",
                "description": "Create multiple CurriculumDetail with the provided details",
                "operationId": "createMultipleCurriculumDetail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/MultipleCurriculumDetail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CurriculumDetail created successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/GetCurriculumDetailsResponse200"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}